Moved appointment video call to separate component
